LayoutTests/fast/viewport/ios/use-minimum-device-width-for-page-without-viewport-meta-expected.txt

setMinimumEffectiveWidth(640.00)
window size: [640, 1096]
square size: [64, 110]
zoom scale: 0.50

setMinimumEffectiveWidth(768.00)
window size: [768, 1315]
square size: [77, 132]
zoom scale: 0.42

setMinimumEffectiveWidth(834.00)
window size: [834, 1428]
square size: [83, 143]
zoom scale: 0.38

setMinimumEffectiveWidth(980.00)
window size: [980, 1678]
square size: [98, 168]
zoom scale: 0.33

setMinimumEffectiveWidth(1024.00)
window size: [1024, 1754]
square size: [102, 175]
zoom scale: 0.31

setMinimumEffectiveWidth(1112.00)
window size: [1112, 1904]
square size: [111, 190]
zoom scale: 0.29

setMinimumEffectiveWidth(1280.00)
window size: [1280, 2192]
square size: [128, 219]
zoom scale: 0.25

setMinimumEffectiveWidth(1336.00)
window size: [1336, 2288]
square size: [134, 229]
zoom scale: 0.24
